[Polling] separate tests into one class per file (#51)
